Create alarm_trigger form and drop-down menu
We inserted form_for to create a form. We used f.select to make the drop-down menu. The first string in each array is what is shown in the drop-down menu, the second string shows up in the URL. It has to be written in lower case and without spaces. 
We planned to split the commit, but GitX didn't allow it for some reason.