Add 'event' to expected_events_params in controller
This is a new part of Rails 4 which is in the 'private' section in the controller and adds additional security to the app. Instead of using attributes accessible in the model, these have to be set in the controller. 

Also, when using validates_presence_of, this has to be included in the model AND in the controller now as well.