Add typeahead to Alarm Trigger form
in form_for we added :data => {:provide => "typeahead"} to the title field.
We also added some Javascript at the end and and the beginning. The autofill information is currently coming from an array. We will have to change that in the future and make it come from the expected events table.