Change drop down menu display
When we chose an option from the drop down menu, it was displayed on the table all smooched together due to the second attribute in our arrays. We thought that they had to be written smooched together like this for it to work but apparently this is not the case. We therefore changed them to make them prettier on the table display.