Remove title validation for Alarm Trigger
We did this because we were getting an error message due to Select2. It is no longer 'required' for the user to enter a title, because they're automatically filled in with the drop down. However, as it stands now, the elements from the drop down are not recognized as being the title in our index table. Seeing as this is not a priority (all the data will change and will come from somewhere else later), we decided to leave this for now.