Add expected event id to alarm triggers
We created an expected event id column to our alarm trigger table, meaning that alarm triggers belongs_to expected events. When working with polymorphic databases, it is important to remember that table that belongs_to another will be the one to receive the migration. 
The expected_event_id is, of course, an integer and is also the "foreign key" in the alarm trigger database (because it comes from somewhere else, it is foreign).