Add validates_inclusion_of for action
We're not entirely sure if this is right or if this is what Carsten meant. We think that we have to somehow go into the browser and see if we can change the drop down menu (for email and logger) - and we think that in theory this validation works as a security feature which will stop ppl from externally messing around with our app. However, we're not sure if this works/how to do this/if this is the case. We don#t get an error message though, which is good. TBC when Carsten is zurück.